moving common function and lots report functionality into separate files.
